Fix reference to use 64 bit ints for multiplcation instead of 32 bits

+++ test/tools/reference_simulator/asm_reference_simulator.cpp
@@ -196,7 +196,7 @@ void simulateMIPSHelper(unordered_map<uint32_t, uint32_t> &memory, uint32_t pc, 
                 if (((regs[sReg] >> 31) && !(regs[tReg] >> 31)) || (!(regs[sReg] >> 31) && (regs[tReg] >> 31)))
                 {
                     // signed multiplication
-                    result = static_cast<int32_t>(regs[sReg]) * static_cast<int32_t>(regs[tReg]);
+                    result = (int64_t)static_cast<int32_t>(regs[sReg]) * (int64_t)static_cast<int32_t>(regs[tReg]);
                 }
                 else
                 {
@@ -211,7 +211,7 @@ void simulateMIPSHelper(unordered_map<uint32_t, uint32_t> &memory, uint32_t pc, 
             {
                 uint32_t sReg, tReg;
                 tie(ignore, sReg, tReg, ignore) = decodeArithmeticType(instruction);
-                uint64_t result = regs[sReg] * regs[tReg];
+                uint64_t result = (uint64_t)regs[sReg] * (uint64_t)regs[tReg];
                 lo = (result << 32) >> 32; // lower 32 bits
                 hi = result >> 32;         // higher 32 bits
                 pc++;
